Add expense shift to group detail page from group header component
