+++ src/app/components/home/group-detail/expense/expense.models.ts
@@ -14,7 +14,7 @@ export interface UpsertExpenseRequest {
   description: string;
   totalAmount: number;
   splitType: SplitType;
-  paidByUserId: string;
+  paidById: string;
   splitDetails: SplitDetailDTO[];
 }
